(feat): Add validation and exception handling for LeaveType
Introduced FluentValidation for LeaveType creation, including a validator and repository method to check uniqueness. Added custom BadRequestException and NotFoundException classes, and updated command/query handlers to throw these exceptions for invalid input and missing records. Updated project dependencies to include FluentValidation.

+++ src/Core/HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,6 +1,7 @@
 using System;
 using AutoMapper;
 using HR.LeaveManagement.Application.Contacts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -20,6 +21,8 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
     // Verify that record exists
+    if (leaveTypeToDelete == null)
+      throw new NotFoundException(nameof(LeaveType), request.Id);
 
     // Delete from database
     await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
